Move Path to std::string
The goal of these series of changes is to remove wstring
from ORBIT code base to then use abseil for all string
operations.

Test: cmake --build .

+++ OrbitCore/ScopeTimer.h
@@ -47,7 +47,7 @@ class Timer {
     return ElapsedSeconds();
   }
 
-  static inline int GetCurrentDepthTLS() { return CurrentDepth; }
+  static inline size_t GetCurrentDepthTLS() { return CurrentDepth; }
   static inline void ClearThreadDepthTLS() { CurrentDepth = 0; }
 
   static const int Version = 0;
